+++ development/ts-migration-dashboard/files-to-convert.json
@@ -1531,7 +1531,6 @@
   "ui/pages/settings/contact-list-tab/view-contact/view-contact.container.js",
   "ui/pages/settings/experimental-tab/experimental-tab.component.js",
   "ui/pages/settings/experimental-tab/experimental-tab.component.test.js",
-  "ui/pages/settings/experimental-tab/experimental-tab.constant.js",
   "ui/pages/settings/experimental-tab/experimental-tab.container.js",
   "ui/pages/settings/experimental-tab/index.js",
   "ui/pages/settings/flask/snaps-list-tab/index.js",
@@ -1580,6 +1579,7 @@
   "ui/pages/settings/settings-tab/index.js",
   "ui/pages/settings/settings-tab/settings-tab.component.js",
   "ui/pages/settings/settings-tab/settings-tab.component.test.js",
+  "ui/pages/settings/settings-tab/settings-tab.constant.js",
   "ui/pages/settings/settings-tab/settings-tab.container.js",
   "ui/pages/settings/settings.component.js",
   "ui/pages/settings/settings.component.test.js",
