+++ ui/pages/settings/experimental-tab/experimental-tab.container.js
@@ -5,14 +5,12 @@ import {
   setUseCollectibleDetection,
   setOpenSeaEnabled,
   setEIP1559V2Enabled,
-  setTheme,
   setCustomNetworkListEnabled,
 } from '../../../store/actions';
 import {
   getUseCollectibleDetection,
   getOpenSeaEnabled,
   getEIP1559V2Enabled,
-  getTheme,
   getIsCustomNetworkListEnabled,
 } from '../../../selectors';
 import ExperimentalTab from './experimental-tab.component';
@@ -22,7 +20,6 @@ const mapStateToProps = (state) => {
     useCollectibleDetection: getUseCollectibleDetection(state),
     openSeaEnabled: getOpenSeaEnabled(state),
     eip1559V2Enabled: getEIP1559V2Enabled(state),
-    theme: getTheme(state),
     customNetworkListEnabled: getIsCustomNetworkListEnabled(state),
   };
 };
@@ -33,7 +30,6 @@ const mapDispatchToProps = (dispatch) => {
       dispatch(setUseCollectibleDetection(val)),
     setOpenSeaEnabled: (val) => dispatch(setOpenSeaEnabled(val)),
     setEIP1559V2Enabled: (val) => dispatch(setEIP1559V2Enabled(val)),
-    setTheme: (val) => dispatch(setTheme(val)),
     setCustomNetworkListEnabled: (val) =>
       dispatch(setCustomNetworkListEnabled(val)),
   };
