+++ ui/pages/settings/settings-tab/settings-tab.constant.js
